+++ opaque_keys/edx/tests/test_collection_locators.py
@@ -4,6 +4,7 @@ Tests of LibraryCollectionLocator
 import ddt
 from opaque_keys import InvalidKeyError
 from opaque_keys.edx.tests import LocatorBaseTest
+from opaque_keys.edx.keys import CollectionKey
 from opaque_keys.edx.locator import LibraryCollectionLocator, LibraryLocatorV2
 
 
@@ -32,11 +33,12 @@ class TestLibraryCollectionLocator(LocatorBaseTest):
         lib_key = LibraryLocatorV2(org=org, slug=lib)
         coll_key = LibraryCollectionLocator(lib_key=lib_key, collection_id=code)
         lib_key = coll_key.lib_key
-        self.assertEqual(str(coll_key), "lib-collection:TestX:LibraryX:test-problem-bank")
-        self.assertEqual(coll_key.org, org)
-        self.assertEqual(coll_key.collection_id, code)
-        self.assertEqual(lib_key.org, org)
-        self.assertEqual(lib_key.slug, lib)
+        assert str(coll_key) == "lib-collection:TestX:LibraryX:test-problem-bank"
+        assert coll_key.org == org
+        assert coll_key.collection_id == code
+        assert lib_key.org == org
+        assert lib_key.slug == lib
+        assert isinstance(coll_key, CollectionKey)
 
     def test_coll_key_constructor_bad_ids(self):
         lib_key = LibraryLocatorV2(org="TestX", slug="lib1")
@@ -52,12 +54,15 @@ class TestLibraryCollectionLocator(LocatorBaseTest):
         code = 'test-problem-bank'
         str_key = f"lib-collection:{org}:{lib}:{code}"
         coll_key = LibraryCollectionLocator.from_string(str_key)
+        assert coll_key == CollectionKey.from_string(str_key)
+        assert str(coll_key) == str_key
+        assert coll_key.org == org
+        assert coll_key.collection_id == code
         lib_key = coll_key.lib_key
-        self.assertEqual(str(coll_key), str_key)
-        self.assertEqual(coll_key.org, org)
-        self.assertEqual(coll_key.collection_id, code)
-        self.assertEqual(lib_key.org, org)
-        self.assertEqual(lib_key.slug, lib)
+        assert isinstance(lib_key, LibraryLocatorV2)
+        assert lib_key.org == org
+        assert lib_key.slug == lib
+        assert coll_key.context_key == lib_key
 
     def test_coll_key_invalid_from_string(self):
         with self.assertRaises(InvalidKeyError):